+++ poseidas/src/main/java/de/governikus/eumw/poseidas/server/pki/MasterAndDefectListHandler.java
@@ -15,8 +15,11 @@ import java.net.URISyntaxException;
 import java.net.URL;
 import java.security.GeneralSecurityException;
 import java.security.KeyStore;
+import java.security.SignatureException;
 import java.util.Arrays;
 import java.util.Base64;
+
+import org.bouncycastle.cms.CMSException;
 
 import de.governikus.eumw.poseidas.config.schema.PkiServiceType;
 import de.governikus.eumw.poseidas.eidserver.model.signeddata.MasterList;
@@ -178,12 +181,17 @@ public class MasterAndDefectListHandler extends BerCaRequestHandlerBase
     if (!isLocalZip(masterList))
     {
       CmsSignatureChecker checker = new CmsSignatureChecker(pkiConfig.getMasterListTrustAnchor());
-      if (!checker.checkEnvelopedSignature(masterList, cvcRefId))
+      try
+      {
+        checker.checkEnvelopedSignature(masterList);
+      }
+      catch (SignatureException | CMSException e)
       {
         SNMPDelegate.getInstance()
                     .sendSNMPTrap(OID.MASTERLIST_SIGNATURE_WRONG,
                                   SNMPDelegate.MASTERLIST_SIGNATURE_WRONG + " "
                                                                   + "signature check for master list failed");
+        log.debug("Signature check on master list not successful", e);
         return null;
       }
     }
@@ -196,13 +204,18 @@ public class MasterAndDefectListHandler extends BerCaRequestHandlerBase
     byte[] defectList = wrapper.getDefectList();
     if (!isLocalZip(defectList))
     {
-      CmsSignatureChecker checker = new CmsSignatureChecker(ml);
-      if (!checker.checkEnvelopedSignature(defectList, cvcRefId))
+      CmsSignatureChecker checker = new CmsSignatureChecker(ml.getCertificates());
+      try
+      {
+        checker.checkEnvelopedSignature(defectList);
+      }
+      catch (SignatureException | CMSException e)
       {
         SNMPDelegate.getInstance()
                     .sendSNMPTrap(OID.DEFECTLIST_SIGNATURE_WRONG,
                                   SNMPDelegate.DEFECTLIST_SIGNATURE_WRONG + " "
                                                                   + "signature check for defect list failed");
+        log.debug("Signature check on defect list not successful", e);
         return null;
       }
     }